Started refactoring the TrayIcon into its own service, using DI.
