--- src/Sidekick.Business/Notifications/INotificationService.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace Sidekick.Business.Notifications
-{
-    public interface INotificationService
-    {
-        event EventHandler TrayNotified;
-
-        void NotifyTray(Notification notification);
-    }
-}
