--- src/Sidekick.Business/Notifications/Notification.cs
@@ -1,8 +1,0 @@
-namespace Sidekick.Business.Notifications
-{
-    public class Notification
-    {
-        public string Title { get; set; }
-        public string Text { get; set; }
-    }
-}
