--- src/Sidekick.Business/Notifications/NotificationEvent.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace Sidekick.Business.Notifications
-{
-    public class NotificationEvent : EventArgs
-    {
-        public Notification Notification { get; set; }
-    }
-}
