+++ src/Sidekick.Business/Platforms/IPlatformTrayService.cs
@@ -1,0 +1,8 @@
+namespace Sidekick.Business.Platforms
+{
+    public interface IPlatformTrayService
+    {
+        void UpdateLeagues();
+        void SendNotification(string title, string text);
+    }
+}
